requirements.txt
fastapi
boto3
uvicorn
python-jose
pydantic
mangum
